Rename PushNotificationManager to PushNotificationHandler
